requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
cryptography
passlib
python-jose
bcrypt